Refactor ValidateRequestDataMixin to mixin module
Add responses module
